Changed elbow bend constraint

+++ ExerciseMotionTracker/Fly.swift
@@ -10,24 +10,27 @@ import Foundation
 
 final public class Fly: ExerciseMonitor
 {
-    let EXERCISE_RANGE: Float = (Float(32) * Float(M_PI)) / Float(180)
-    var MAX_FOREARM_ANGLE: Float = (Float(86) * Float(M_PI)) / Float(180)
-    var BENT_ELBOW_CONSTRAINT: Float = (Float(25) * Float(M_PI) / Float(180))
+    let EXERCISE_RANGE: Float = (Float(32) * Float(M_PI)) / Float(180) // range is 32 degrees
+    var MAX_FOREARM_ANGLE: Float = (Float(86) * Float(M_PI)) / Float(180) // max angle is 86 degrees
+    var BENT_ELBOW_CONSTRAINT: Float = (Float(20) * Float(M_PI) / Float(180)) // max elbow bend is 25 degrees
     
-    // returns percent complete of a bicep curl
+    // returns percent complete of a fly exercise
     override public func getPercentComplete() -> Int? {
         
         var percentComplete: Int
         var percentCompleteFloat: Float
         
+        // get sensor orientations
         let rightBicep = skeleton.bodyJoints["rightBicep"]
         let rightForearm = skeleton.bodyJoints["rightForearm"]
         let leftBicep = skeleton.bodyJoints["leftBicep"]
         let leftForearm = skeleton.bodyJoints["leftForearm"]
         
+        // determine angle between forearm sensors and angle between bicep sensors
         let forearmAngle = rightForearm?.getAngleBetweenJoints(otherJoint: leftForearm!)
         let bicepAngle = rightBicep?.getAngleBetweenJoints(otherJoint: leftBicep!)
         
+        // constrain range of angle
         if (forearmAngle! > MAX_FOREARM_ANGLE)
         {
             percentComplete = 0;
